backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
httpx==0.28.1
beautifulsoup4==4.12.3
python-docx==1.1.2
pdfplumber==0.11.5
apscheduler==3.11.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
